P5_03_codesource/requirements.txt
mysql_connector
requests
tqdm